refactor: update import paths for AudioPool and remove unused pool.ts file
