+++ src/context/audio.tsx
@@ -2,7 +2,7 @@ import type React from 'react'
 import { createContext, useContext, useEffect, useRef, useState } from 'react'
 
 import { useSettings } from '@/context/setting'
-import AudioPool from '@/utils/audio/pool'
+import { AudioPool } from '@/utils/audio/audioPool'
 
 interface AudioContextType {
   playSound: (
