+++ src/utils/sound/index.ts
@@ -1,7 +1,7 @@
 import HotkeyManager from '@/lib/system/hotkeys'
 import type { Config } from '@/types/config'
 import type { SoundData } from '@/types/sound'
-import AudioPool from '@/utils/audio/pool'
+import { AudioPool } from '@/utils/audio/audioPool'
 import { getElement } from '@/utils/getElement'
 import { generateId } from '@/utils/sound/generateId'
 import { getMusic } from '@/utils/sound/getMusic'
